[bug] setLoading auto-resets to false
I think the identity of setLoading is different in AppContext and
RouteContext, but they set the same value somehow. The values are then
one step behind of each other: whenever you call setLoading from
RouteContext (value 1), it sets the new value, but the setLoading from
AppContext is executed and re-sets the value to the previous value
(value 0). Then, when setLoading is called againg from RouteContext, it
sets the value to (value 2), but setValue from AppContext re-sets the
value to (value 1). Is just an hypothesis.

This is the application setting the first loading value:

  loading changed to false
  loading.tsx:12 loading changed to true

This is the <GameRoute /> setting the loading to true while the
useGameStatusQuery resolves (enforcing a 3 seconds delay so that
I can actually see what's going on):

  route-game.tsx:34 wait 3 seconds...
  loading.tsx:12 loading changed to true
  loading.tsx:12 loading changed to false

This is the data loading:

  route-game.tsx:53 {
    "__typename": "GamePlaying",
    "id": "7a7acdd7-f061-4478-b4ad-81a796a490f4",
    "me": "O",
    "next": "O"
  }

This is called last because setLoading is called in an useEffect hook:

  route-game.tsx:34 wait 3 seconds...
  loading.tsx:12 loading changed to true
  loading.tsx:12 loading changed to false